+++ DITesting/Program.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 
 namespace DITesting;
 
@@ -34,84 +33,6 @@ internal class Program
     {
         new AppBuilder().Build().Run();
     }
-}
-
-internal class AppBuilder
-{
-    private readonly IServiceCollection _services = new ServiceCollection();
-    private IConfiguration? _configuration;
-
-    private App App;
-
-    public AppBuilder Configure(Action<IConfigurationBuilder> configure)
-    {
-        var configBuilder = new ConfigurationBuilder();
-        configure(configBuilder);
-
-        _configuration = configBuilder.Build();
-        _services.AddSingleton<IConfiguration>(_configuration);
-
-        return this;
-    }
-
-    public AppBuilder ConfigureServices(Action<IServiceCollection> configure)
-    {
-        configure(_services);
-        return this;
-    }
-
-    public App Build()
-    {
-        if (_configuration == null)
-            throw new InvalidOperationException("Configuration must be set before building the app.");
-
-        App = new App(_services.BuildServiceProvider(), _configuration);
-        return App;
-    }
-}
-
-internal class App
-{
-    private static IConfiguration? _configuration;
-    private readonly IServiceProvider _serviceProvider;
-
-    public App(IServiceProvider serviceProvider, IConfiguration configuration)
-    {
-        _serviceProvider = serviceProvider;
-        _configuration = configuration;
-    }
-
-    public void Run(CancellationToken cancellationToken)
-    {
-        try
-        {
-            Console.WriteLine("app started. Resolving services.");
-
-            var scopeFactory = _serviceProvider.GetRequiredService<IServiceScopeFactory>();
-            using var scope = scopeFactory.CreateScope();
-
-            // This is just an example of how to use the scopes. In App we should always use '_serviceProvider'.
-            var system = scope.ServiceProvider.GetRequiredService<FrameSystemInterface>();
-
-            Console.WriteLine("Required services resolved.");
-            Console.WriteLine("Entering main loop.");
-
-            bool running = !cancellationToken.IsCancellationRequested;
-            while (running)
-            {
-
-                system.Update();
-
-                Thread.Sleep(1000); // Simulate frame delay
-            }
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Exception occurred: {ex}");
-        }
-    }
-
-    public void Run() => Run(CancellationToken.None);
 }
 
 internal interface FrameSystemInterface
