+++ SharpEngine.Core/SharpEngine.Core.csproj
@@ -27,6 +27,9 @@
     
     <PackageReference Include="StbImageSharp" Version="2.30.15" />
     <PackageReference Include="System.Text.Json" Version="8.0.5" />
+
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="10.0.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="10.0.1" />
   </ItemGroup>
 
   <ItemGroup>
